chore(icons): build cjs bundle
