+++ packages/design-system-icons/package.json
@@ -38,7 +38,10 @@
   "scripts": {
     "test": "jest",
     "dev": "node ./builder.mjs && webpack --watch & yarn storybook",
-    "build": "node ./builder.mjs && tsc -p tsconfig.build.json && webpack",
+    "build:prepare": "node ./builder.mjs && tsc -p tsconfig.build.json",
+    "build:esm": "webpack -c ./esm.config.js",
+    "build:cjs": "webpack -c ./cjs.config.js",
+    "build": "yarn build:prepare && yarn build:esm && yarn build:cjs",
     "storybook": "start-storybook -p 6006",
     "build-storybook": "build-storybook --docs && npx -y storybook extract "
   },
